+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Fixed inaccuracies in weapon compendium.
 - Redesigned various item sheets.
+- Item text editor uses prosemirror.
 
 ## [0.2.22] - 2022-10-19
 ### Added